pyproject.toml
[project]
name = "movai"
version = "0.1.0"
description = "Movie AI chatbot powered by Cohere and TMDB"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-mail>=0.10",
    "cohere>=5.9",
    "requests>=2.32",
    "python-dotenv>=1.0",
    "mixpanel>=4.10",
    "waitress>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3",
]

[project.scripts]
movai = "movai.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
